Refactor useEmployment hook as a util/hook.js
- move useEmployment to utils
- refactor export in lib.js
